+++ src/drawContainer.js
@@ -136,9 +136,7 @@ const draw = async (sectors, config = {}, outputFilename) => {
   }
 
   if (fn) {
-    console.log('starting function');
     fn(ctx, screenCoord, can);
-    console.log('done with function');
   }
 
   const stream = can.createPNGStream();
@@ -149,10 +147,7 @@ const draw = async (sectors, config = {}, outputFilename) => {
   const p = new Promise((d) => {
     done = d;
   });
-  const dt = Date.now();
   writeStream.on('close', () => {
-    console.log('writeStream done writing', outputFilename);
-    console.log(((Date.now() - dt) / 1000).toFixed(1), 'seconds');
     done();
   });
   return p;