+++ src/index.js
@@ -1,7 +1,11 @@
 import HexRegion from './HexRegion';
 import randomFor from './randomFor';
 import Universe from './sectors/Universe';
+import {
+  TEN, HUN, K, K10, K100, M, M10, M100, B, B10, B100, T,
+} from './constants';
 
 export default {
   HexRegion, randomFor, Universe,
+  TEN, HUN, K, K10, K100, M, M10, M100, B, B10, B100, T,
 };