+++ test/HexRegion.spec.js
@@ -63,7 +63,7 @@ tap.test(p.name, (suite) => {
       div.end();
     });
 
-    hr.skip('draw divided', async (dd) => {
+    hr.test('draw divided', async (dd) => {
       const baseRegion = new HexRegion({ diameter: 110 });
 
       baseRegion.divide(4);
@@ -132,7 +132,7 @@ tap.test(p.name, (suite) => {
       dd.end();
     });
 
-    hr.skip('draw sub divided', async (dd) => {
+    hr.test('draw sub divided', async (dd) => {
       const baseRegion = new HexRegion({ diameter: 110 });
 
       baseRegion.divide(5);
@@ -195,82 +195,23 @@ tap.test(p.name, (suite) => {
     hr.end();
   });
 
-  suite.skip('noise on hex', async (noh) => {
-    const baseRegion = new HexRegion({ diameter: 40 });
+  /**
+   * skip this test -- its got no asserts
+   */
+  suite.skip('relationship of single division and nested division', (md) => {
+    const singleDiv = new HexRegion({ diameter: 100 });
+    singleDiv.divide(16);
 
-    baseRegion.divide(300);
-    const bounds = baseRegion.toBox();
+    const nestedDiv = new HexRegion({ diameter: 100 });
+    nestedDiv.divide(4);
+    nestedDiv.forEach((ns) => {
+      ns.divide(4);
+      ns.forEach((nss) => {
+        console.log('...subsector ', nss.id, 'center = ', nss.center.clone().round());
+      });
+    });
 
-    const sn = new Noise(randomFor('test'));
-    const sn2 = new Noise(randomFor('everything'));
-    const sn3 = new Noise(randomFor('life'));
-    const sn4 = new Noise(randomFor('the universe'));
-
-    await drawHexes({
-      min_x: bounds.min.x * 1.5,
-      min_y: bounds.min.y * 1.5,
-      max_x: bounds.max.x * 1.5,
-      max_y: bounds.max.y * 1.5,
-      visual_scale: 100,
-      fn(ctx, screenCoord) {
-        const drawHex = (hex, color = 'black', width = 1, fill = false) => {
-          ctx.lineWidth = width;
-          ctx.strokeStyle = color;
-
-          const corners = hex.corners().map(screenCoord);
-
-          corners.push(corners[0]);
-
-          ctx.beginPath();
-          const p = corners[0];
-          ctx.moveTo(p.x, p.y);
-          corners.forEach((p2) => ctx.lineTo(p2.x, p2.y));
-          ctx.closePath();
-          if (fill) {
-            ctx.fillStyle = fill;
-            ctx.fill();
-          }
-          ctx.stroke();
-        };
-
-        drawHex(baseRegion, 'black', 2);
-
-        baseRegion.forEach((child) => {
-          const center = child.center.clone().multiplyScalar(12 / baseRegion.diameter);
-          const center2 = child.center.clone().multiplyScalar(30 / baseRegion.diameter);
-          const center3 = child.center.clone().multiplyScalar(200 / baseRegion.diameter);
-          const center4 = child.center.clone().multiplyScalar(50 / baseRegion.diameter);
-          const shade = 1 - Math.abs(sn.noise2D(center.x, center.y));
-          const shade2 = 1 - Math.abs(sn2.noise2D(center2.x, center2.y));
-          const shade3 = sn3.noise2D(center3.x, center3.y);
-          const shade4 = 1 - Math.abs(sn4.noise2D(center4.x, center4.y));
-          const c = clamp(Math.floor(((shade * 0.4 + shade4 * 0.25 + shade2 * 0.15 + shade3 * 0.1) ** 1.25) * 255),
-            0, 255);
-          drawHex(child, `rgb(${c},${c},${c}`, 1, `rgb(${c},${c},${c}`);
-        });
-      },
-
-    }, 'noise on hex');
-
-    noh.end();
-  });
-
-  suite.skip('math', (m) => {
-    const h = new Hexes({ pointy: true, scale: 1000 });
-
-    const origin = new CubeCoord(0, 0);
-
-    console.log('100 scale hex points: ', h.corners(origin));
-
-    console.log('corner to corner distance: ', h.corners(origin)[0].distanceTo(h.corners(origin)[3]));
-
-    const hf = new Hexes({ pointy: false, scale: 100 });
-
-    console.log('100 scale hex, flat points: ', hf.corners(origin));
-
-    console.log('corner to corner(flat) distance: ', hf.corners(origin)[0].distanceTo(hf.corners(origin)[3]));
-
-    m.end();
+    md.end();
   });
 
   suite.end();